pyproject.toml
[tool.poetry]
name = "mrdamian"
version = "0.1.0"
description = ""
authors = ["Yuniru Yuni <yuniruyuni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
twitchio = "^2.5.0"
deepl = "^1.11.0"
pywin32 = "^305"
numpy = "^1.23.1"
soundcard = "^0.4.2"
whisper = {git = "https://github.com/openai/whisper.git"}
torch = {url = "https://download.pytorch.org/whl/cu113/torch-1.10.1%2Bcu113-cp39-cp39-win_amd64.whl"}
transformers = "^4.24.0"
ffmpeg = "^1.4"
tqdm = "^4.64.1"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
flake8 = "^6.0.0"
black = "^22.12.0"
mypy = "^0.991"
nox = "^2022.11.21"
nox-poetry = "^1.0.2"
pyproject-flake8 = "^6.0.0.post1"

[tool.flake8]
exclude = [
    ".venv",
    ".nox"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
